Don't allow block settings outside of extent and OffsetExtent -> OffsetTransform
